노마드코더 - 그림판만들기 (공부)2
Javascipt codes were added.

There are these functions, like drawing on the canvas with the pen that is able to change its size and colors and changing the whole canvas with in 9 colors.

I will add one more function that is to convert to image file from the work space (paint board).